+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## HEAD
 
-* Fix for broken cookie support
+* fix: broken cookie support
 * If html report is generated, convert it automatically to pdf; can be disabled with `--no-pdf`
 * dependency: update to puppeteer 15.2.0 (Chrome/93.0.4577.0 -> Chrome/104.0.5109.0)
+* fix: unresolvable paths in website-evidence-reporter
+
 
 ## 2.0.0 / 2022-06-30